Get bounce working, and some cleanup
Implemented a bunch of functions and stubbed others to get bounce compiling and hopefully working.

While in the code, I cleaned up tracing a little bit.  I also took
functions that called other functions and and made a base function
that both of those would call instead.  That way there won't be
misleading entries in the trace.

* Make "if(trace_functions)" into a pair of macros: TRACEF and TRACE, mostly just to look a little better.
* Implement {bgn,end}{line,closedline,swaptmesh}, v4f, c3i
* stub czclear(), zfunction()
* Implement gconfig() but who knows...
* Insert draw(), drawi(), movei(), move(), pclos(), pdri(), pmvi(), poly()  into display lists