Double-buffer the command stream
If "double_buffer_commands" is set to 0, commands are written directly to the GPU memory buffer for commands.  (Therefore no additional commands can be safely written to the buffer until the GPU has finished processing the buffer.) "swap" starts the GPU and waits for the GPU to finish before returning.

If "double_buffer_commands" is set to 1, commands are written directly to a separate "staging" memory buffer.  "swap" waits until any previous GPU work is complete, meaning the GPU buffer can be written safely, then copies the staging buffer to the GPU buffer and starts GPU work.

N.B.  By default now set double_buffer_commands to 1.  It is a const variable meant to be compile-time-only.  Before converting this variable to settable at run-time, there must be a synchronization step before changing the variable.