Break up long thin triangles. For issue #17.
Long thin (diagonal) triangles waste a lot of time in the
rasterizer because we consider every point in the bounding
box. With this change, if the ratio of triangle pixels
to bounding box pixels is low, we tesselate the triangle.

This halves the number of bounding box pixels (1.8m to 0.9m)
at the slowest frame of buttonfly. (The bevel of the button
flying toward the camera is made of long thin diagonal triangles.)

This should shave about 8ms off the rendering time (18m down
to 9.3ms), but it only shaves a few milliseconds because we
generate an extra 4000 triangles, and the per-triangle
overhead in the rasterizer eats up about 4.7ms for those.
That's for another day.